requirements.txt
accelerate
datasets
dotenv
einops
# for gemini api
google-genai
gradio
huggingface_hub
imageio
ipdb
json_repair
matplotlib
nltk
numpy
omegaconf>=2.4.0.dev4
openai
opencv-python>=4.7.0.72
openpyxl
pandas
pillow
portalocker
protobuf
pylatexenc==2.10
python-dotenv
qwen_vl_utils
requests
rich
scikit-learn
sentencepiece
setuptools
sty
sympy
tabulate
tiktoken
timeout-decorator
timm
# torch
# torchvision
tqdm
transformers
typing_extensions
validators
xlsxwriter
scikit-learn
datasets
apted>=1.0.3
colormath>=3.0.0
decord>=0.6.0
distance>=0.1.3
lxml>=6.0.2
pdf2image>=1.17.0
zss>=1.2.0
polygon3>=3.0.9.1
levenshtein>=0.27.1
jieba>=0.42.1
editdistance>=0.8.1
anls>=0.0.2
antlr4-python3-runtime==4.11.1
# For SArena
torchmetrics
scikit-image
lpips
openai-clip
cairosvg
# For UniSVG
sentence_transformers
bert_score
